[r2dbe_vdif_lib] Added initial VDIF headers
Important: the bug reported in the previous commit of the 0th
packet having a size 64 bits short turns out to be specific
to the tengbe block (reset before streaming data), and it
is confirmed that the version of the vdif block concerned
here in this commit and in the previous commit are not
producing this error.

The difference between this and the previous commit is
therefore simply the addition of meaningful VDIF headers.
This is just my initial vdif header set up, and may be adjusted
in future commits.  Currently the header is set:

l = logic, either software registers, counters, or both

w0|0|0|                             l                          |
w1|0|0|   l    |                       l                       |
w2|  0  |   0  |                      1024                     |
w3|0|   1   |         l          |               l             |
w4|     2      |                     678629            |l| l |l|
w5|               l              |               l             |
w6|                              l                             |
w7|                              l                             |

Each word being 32 bits, they are stacked into a single 64 bit
word for transmission.

Currently, with no setting of any VDIF software registers, I
can run VDIF.py and the capture packet script receives:
   0: 0x00000008
   1: 0x0000C672  <-- increments once per pkt, rst on pps
   2: 0x00000400  <-- static, 1024
   3: 0x04000000  <-- static, 1 in 26th bit
   4: 0x02A5AE50  <-- static
   5: 0x00000000  <-- static
   6: 0x000EF03F  <-- increments once per pkt, forever
   7: 0x00000000  <-- MSW of ctr above, so will change eventually