[software] Created a software sub-directory and added a basic VDIF receiver.
This script will capture N number of UDP packets and print out the entire
VDIF header as well as parts of the beginning and end of the data. To run
use the following command:

$ python ./software/vdif_recv.py <N>

Where <N> should be replaced by the numer of packets you'd like to grab.
Note: this has the IP hard-coded and will only work of course if you have
a 10 GbE card connected directly to the ROACH2 running the R2DBE bitcode.