[casper_files/r2dbe_vdif_lib.slx] synchronized pps through header generation
The PPS signal in the top level file is currently not internally
generated, and because of this, counters in the headers that were
enabled by the pps rather than the rising edge of the pps were
incrementing too quickly (the pps signal is high for many samples).

Rurik is currently working on the internally generated 1 pps signal
which will be high for a single sample.  I went into the VDIF block
and removed any rising edge blocks after the pps as I had done this
inconsistently, and I have also ensured that any headers that require
the pps to trigger an enable or reset are synchronized to execute
at the same time.