[r2dbe_adc_lib] The quantize block now interprets data as signed with binary point.
Before it would interpret its input data as unsigned 8-bits with no binary point,
then multiply by the gain factor which was unsigned 16-bits with a 12-bit binary
point, until finally re-interpreting the output as unsigned 2-bit no binary point.
This was non-ideal and resulted in incorrect 2-bit statistics.

This commit changes this scheme to use signed and N-1 binary point data for both
the 8-bit input and the 2-bit output, while changing the programmable gain to an
unsigned 16-bit, 8-bit binary point value, allowing for a range of gains of
1/256 <= gain < 256.