[r2dbe_top, r2dbe_vdif_lib] Added a data reorder block to the front of the vdif block
I've added a block to reorder the 2-bit samples to match the VDIF
spec.  I've added a software register at the top level of the VDIF
block called:
    reorder_2b_samps
which, when set to 1, will reorder the 2 bit data such that the
oldest sample occupies the 2 LSBs of the 32-bit word.  The ADC
natively supplies the data such that the oldest sample is in
the highest bits.

I've also moved the little endian swap software register to the
top level of the VDIF subsystem for consistency.

This was committed and tested with Rurik's software, the bitcode
is:
    r2dbe_top_2014_Aug_08_1508.bof.gz