[VDIF_test.py] updating with changes to Westford setup
The installation and setup of the Westford Mark6s for the
SPT lead to multiple changes made locally that were not committed.
I am adjusting these on Mark6-4015 and pushing, verified that it
works well for Mark6-4016, currently set up with an r2dbe at SAO.

Additionally, it was found to be vitally important that the "enable"
signals are not set until the pps arm has finished, at least 1 second.
Otherwise, the PPS signal generated on board may come up with one phase
and then jump when the PPS is re-armed.

Unfortunately, this causes the "start of frame" signal and the PPS
to now be misaligned, which can cause a 4 us offset in the data as
well as data frame headers that do not reset at the correct time.

these data headers result in frame 0 truly being labeled frame
125,000.