[software/alc.py] adding automatic level setting for 2b quant
alc.py can be run after VDIF_test.py.  It will connect to a
programmed and initialized R2DBE, and grab snapshots of the 8-bit
data from if0.  The data is then sorted, and then the appropriate
threshold value is found by investigating the value of data 16% of
the way through the vector from both ends.

the average of these two estimates is found and then the correct
register in the R2DBE bitcode is set to this value.

The same is then done for if1.

This code does not need to capture data on the ports, no need to
run as root, and will work even if the 10 GbE cables entering the
from the R2DBE to the Mark6 are unplugged entirely.