[software/vdif.py] Hacked a quick way of viewing the pps offset buried in W5
vdif.py was modified to produce a PSN and the W5 PPS offset in
FPGA clock cycles.  This is currently not done properly because
I need to learn more about python programming and classes, but its
a super quick fix that will allow the south pole telescope to look
at the offset buried in the vdif files if useful for monitoring.