[swarm_2_comp_.._Feb_17...bof] adding boffiles from Feb 17
Most importantly is 1841, which includes a bug fix.  Previous recent
bitcodes do not perform the little endian swap.  This bitcode adds
it back in.

The other bitcode from that day was prior to this fix, but is
again coming along for posterity.

Both these bitcodes made a change to the fake data packets in that
instead of producing data for 256 clocks and off for one, it prod-
uces data for 256 consecutive clocks ever 320 clocks.

The 256 on, off for 1,
while representing continuous reads of b-engine packets
from the 10 GbE port, is actually not the situation we would be in.
The data rate associated with that is nearly 9.9 GbE into the
swarm dbe, which is higher than the 8 we expect, and in addition
does not produce an integer number of packets per second.  Because
of the way I am currently doing data frame counting (based on the
a division of the sync pulse, not the incoming packet itself), we
could have vdif packets with identical headers (data frame number
the same), or the data frame might skip by more than 1, and the
Mark6 may stuff a packet in when that is not intentional.

The 256 on, 64 off (total period of 320) is nearly 8 Gbps (similar
to our actual situation), and represents latency that we would
expect to see, although it may manifest less evenly distributed.