[swarm...slx, swarm...bof] change back to 257 clocks from 320 clocks per pkt
The last boffile/model file had 2 changes: 1) bug-fix to add little
endian byte-swap, and 2) added latency in the packet stream such
that packets were not continuuous.

Turns out that the little endian swap works well, but the latency
caused problems that still need to be debugged.  I removed the
latency such that this code shows that the little endian swap
is performed correctly.