[swarm...bof, swarm...slx] bug-fix for data frame and vtp count
in the last compile, the psn and data frame number would stall at
the same value every few clocks and then skip to catch up.  this
caused the mark6 to stuff packets in between when the data frame
number skips.

There were several reasons for this.  First the header values
were written to their FIFO on a subdivision of a second, not related
to the packet itself.  So even though the value incremented according
to packet, it was written to the FIFO at off-times, causing wrong
values to be put in the header.

Additionally, the data frame value starts at 1 not 0.  Both of
these issues should be fixed in this version.