[swarm...bof, swarm...slx] fixed get_pkts_from_tengbe bug
Many simulations of the fake packetier show that the data sent
off the tengbe is reliable.  this boffile jumps straight to using
the 10gbe rx code, and fixes a bug in how the packets are
interpretted from the tengbe receiver.  ready for testing on the
mountain here on Thursday afternoon Feb 26.