[swarm...bof, swarm...slx] added network buffer to 10gbe design
Added a network buffer to take non-consecutive data words from
the 10 gbe receiver and buffer them to deliver a consecutive packet
to the rest of the bitcode.  I think the lack of this buffering
and the assumption that data would stream out in one consecutive
packet from the 10 gbe lead to the long and varying packet lengths
seen in testing.