[swarm...slx,swarm...bof] fake packets, VDIF fixed
The data buffers responsible for the top and bottom of the 64 bit
words had an error which explains why we see the top data correct
but the bottom data incorrect.  Using swarm_w_comp_diagnose.slx
on my virtual machines shows this error and that it is now fixed.

The valid signal i was sending into the bottom data stayed on
long after data valid stayed off.  Thus data was being written
when it shouldnt have been.