[swarm...bof, swarm...slx] this bitcode ties all 4 outputs to 1 input
So that we can test the full rate into the Mark6 in the lab where
we only have 2 roaches that can send data.

The previous compile has an error where the beng header is off by
one packet.  The result after testing was that the bengine mislabels
the data as coming from the other fid and a different channel id.
this resulted in frequency placement being off by 56 channels.

This compile still has this error, but will be helpful for testing.