Add maser monitoring tools
Typically drift-monitor.sh should be run in the background as root. This
script periodically grabs packets from the ethernet interface, stores it
in /tmp, and then runs get_drift_from_vdif.py to extract the time-offset
data. The current implementation grabs a packet of data every 64 seconds.

For each measurement a log entry is made in the file
/var/log/r2dbe/maser-drift.log. The path should exist, and the file is
created if it does not exist. The log file is in CSV format with
header lines that note the creation date of the logfile and the meaning
of the different columns. The file grows ~16bytes per entry, or ~1MiB per
48days.

plot_maser_drift.py will interpret the logfile, plot the drift over time
since the first entry in the logfile, fit a quadratic curve to the data,
and print some parameters to stdout.