[plot_hist.py] will take vdif file as arg, will plot hist stats
histograms are presented as a colormap, by state and individual
channel so we can see how the state counts change across the
16384 frequency channels in swarm

this is presented for each of the 2 phased sums

additionally, we present an average histogram across channels
specified by the user in the -gl (guardband low) and -gh
(guardband high) flags.  Thus you can look at the histogram
for a specific channel, or a "valid" range of channels.

can also retrieve the current threshold values.

can be used in a script that uses grab_consec.sh to get pkts
from each of the 4 streams incoming to the Mark6, send them
all into this plot_hist.py function (takes multiple files in),
we view the analysis, then close and clean up the data