[plot_hist_corrs,get_data_plot_hist] add scripts for thresh diagnosis
in addition to plot_hist.py found in previous commits, it has been
extended to plot_hist_corrs.py which will plot the histograms as
before, but add a second plot of the auto and cross corrs of the
two phase sums.

additionally included in this commit is the bash script
get_data_plot_hist.sh, which grabs 2048 packets from each ethernet
port, runs plot_hist_corrs.py with some basic flags/values, then
deletes the recordings.  Thus there is a single button to push to
take the pulse of the SDBE data, however it takes roughly 45 seconds
to produce the plots.