[plot_hist_corrs,get_data_plot_hist] bugfix plot_hist_corrs.py, changed channels preferred in get_data_plot_hist
Edited plot_hist_corrs.py so that the cross-corr plots as 10*log10
of the MAGNITUDE of the samples, not the raw complex samples as
incorrectly done previously.

get_data_plot_hist was changed so that by default it looks at a wider
range of channels.  this is just a preference for usefulness, and
is how the script was used this morning to collect test data on
3c84 and set the thresholds.