[uut: vdif_hdr] added model+script for isolated testing of hdr sub-block
Specifically looking at timing errors.  Use matlab workspace
variables set in the uut_vdif_hdr_script.m file to create inputs.

Made data 0 of each second a 1, all other data 0 so we can see data
makes it out of the block with the same latency as everything else.

Have an enable signal that executes some time after start.

pps signal is shortened to 8192 clocks, so that frame size
staying the same as the original block (2048 clocks) we get
an even 4 packets (frames) per second, and can see a few
seconds worth of data in the 40,000 seconds of simulation.

fnum correctly resets and is 0 when pps increments the seconds
since the reference epoch.  These values need to be correct
when the "frame out" signal is valid high, once at the start
of every packet.