[uut_vdif_hdr] modified internals of hdr block so sw regs set by matlab in sim
I've modified the internals of the hdr block just so that the
software registers for reset and sec_ref_ep in W0 are set in
simulation from the Matlab workspace.  That means this hdr
block CANNOT be copied as is into the r2dbe_top.slx file, however
none of the behavior of the block has been altered.

Tests show that this block is functioning as expected.  Small
windows occur where off-by-1s errors could happen, but with
software we can shrink these windows even further (with point
to point and with software as of July 2015 1/100 chance of
off-by-1s error).  These off-by-1 errors are detectable with
the vv diagnostic script, so all sites will have found these.

Currently the fix if this off-by-1 error occurs is that you can
re-run the r2dbe_start script, or you can simply execute the one
katcp command to adjust the sec_ref_ep software register to remove
the 1s time offset.