[r2dbe_start] shrink window of off-by-1s error from 1/100 to 1/200
There is a likelihood when setting W0 of the vdif hdr that you
get an off-by-1s error in the timing.  This occurs because the
computer calculates the current number of seconds since the epoch,
and will eventually send this constant to a software register on
the r2dbe.  Once this number is determined, however, we must
reset a counter that increments every second.  Thus we have
an entry in the VDIF header that is the constant seconds since
the reference epoch as determined once by the computer, added
to a counter that increments every pps.

How off-by-1s occurs:
If we calculate the constant secs since the ref epoch JUST before
the change of a second, the reset signal might get there after this
change of second, thus making the time stamp in the r2dbe 1 second
slow of real time.

With a point to point connection, it takes roughly 10 ms from
determining the constant to sending the reset signal, which means
there is a 1 in 100 chance that one or both data streams end up
in this off-by-1s state.  Those are small odds, and the vv diagnostic
script that all site operators run will indicate the off-by-1s.

To fix, simply re-execute the r2dbe_start.py script, or connect
to the r2dbe (ipython, perhaps) and execute a katcp command to
fix the constant in the software register to remove the offset.

Reducing the likelihood of this:
Additionally, this update reduces the time window that makes the
r2dbe susceptible to this sort of offset by allowing the reset commands
to start high, and putting as few commands as possible between the
calculation of the constant seconds since the reference epoch and setting
the resets low.  Each command takes roughly 2.5 ms, so thus the window
is now only 5ms long.  Same fix as above if this off-by-1s occurs in the
field.