[uut_vdif_serialize_hdr] adding test bed for serialize_hdr block in vdif pipeline
this block simply takes the input header signals that are in 8 in
parallel and serializes them, such that w0 is valid at frame out

no latency so frame_in signal is same as frame_out.

data is not delayed relative to other signals.

hdr data is successfully parallelized, and timing signals in
particular (secs since epoch, frame number) were checked and
are clearly preserved in new parallelized and valid format