[uut_vdif_hdr_script.m] fixed data0 pulse in matlab script
data 0 was actually not being written to the simulation as I had
expected.  I have since fixed this, and will add these changes
to the other units under test.

It looks as though data15 was actually the sample with the pulse,
and then when plotting the data streams there was an error as well.

Fixing these bugs results in the same conclusions in terms of data,
latency, and any delay relative to pps (determined there was none).