[uut_vdif_mk_64] data0 bug fix, was not handling output data right
the data0 bug fix from previous commits was migrated here, fixed,
works as expected.

In terms of the model file and script, I was still reading the
output data as if it were 32 bits, even though now its 64.

Ive fixed that and changed the code to handle this, added some
figures to negotiate the results.  block works as expected, latency
of 2 clocks, no timing errors.