[uut_sdbe] Files added for testing the SDBE pipeline
uut_sdbe_pipeline.slx is a model file with all kinds of matlab sinks
to/from the workspace

uut_sdbe_script.m will prepare input variables, run the simulation,
and plot and analyze some of the results.  Its a bit messy, but
it works.

Other .m files are simple diagnostics needed for the script.