[uut_vdif] a testbed for the entire pipeline
turns out it doesnt take too long to simulate things

the uut_vdif.slx script and the uut_vdif_script.m simulate the
entire pipeline.  There are several checks at the end of the script
that allow you to view that the vdif header data (with regards to
timing signals) are right.

The script calls and simulates the model file, so no need for
breakpoints unless they help your workflow.

the user can specify the timing of the input signals including
the enable, the pps, the counter resets.