[r2dbe_start.py] added a wait to reduce off-by-1s errors to 0
If Mark6 is properly synced to NTP, then we can avoid the
off-by-1s scenario by having the r2dbe_start.py script
wait until the middle of second before calculating the time
and resetting the counter on the r2.  If a pps occurs between
these operations (time to complete on point to point connection
is on the order of 10ms), then we can have an off-by-1s error.

By forcing this to happen in the middle of a second, we remove
the chance that a pps can occur between the two.