Add persistent quadrant settings to startup script
The startup script swarm_dbe.py now accepts a reduced set of
parameters similar to r2dbe_start.py. The quadrant is specified as
the only, and mandatory, positional argument with per-quadrant
settings hard coded in the script. The settings are standard through
the rest of SWARM, and may need some central configuration in the
future.

General clean-up of the startup script too.