[software/r2dbe_start] Add explicit MAC IP config
The startup script can now also use provided MAC / IP addresses
directly. If the iface option is a comma-separated list it is assumed
to contain the MAC address in 0th position and IP address in 1st, with
the former in colon-sparated hexadecimal bytes (aa:bb:cc:dd:ee:ff) and
the latter in dot-decimal notation (1.2.3.4). Otherwise the value of
the iface option is assumed to be the receiving network interface from
which MAC and IP addresses are determined.